cleaning up the architecture
