+++ web/src/components/summary/FileSimilarityHistogram.vue
